Pair up for collaboration and conflicts lessons
I've done this in a previous workshop, and I think it will be clearer
for the students than 'collaborating' between two repositories on a
single computer (I haven't actually tried teaching that, but it seems
counterintuitive).

The downside is that several diagrams are now inapplicable. I've drawn a
couple of new diagrams to go in their place, but I haven't got time to
replace them all, so I've just removed some.